pyproject.toml
[project]
    dependencies = [
        "aiohttp-jinja2>=1.6",
        "aiohttp>=3.8.0",
        "asyncio >=3.4.3",
        "jinja2>=3.1.6",
        "numpy>=2.2.3",
        "open_image_models>=0.2.1",
        "opencv-contrib-python-headless>=4.11.0.86",
        "pillow>=11.1.0",
    ]
    description = "License plate extractor"
    name = "platefinder"
    readme = "README.md"
    requires-python = ">=3.12"
    version = "0.1.0"

[tool.ruff]
    line-length = 160

    [tool.ruff.lint] # Copied these from other projects I have.
        select = [
            "A",
            "ARG",
            "ASYNC",
            "B",
            "BLE",
            "C4",
            "DTZ",
            "E",
            "EM",
            "ERA",
            "FBT",
            "FIX",
            "FURB",
            "INT",
            "N",
            "NPY",
            "PERF",
            "PIE",
            "PYI",
            "Q",
            "RET",
            "RUF",
            "S",
            "SIM",
            "SLOT",
            "TRY",
            "UP",
            "W",
        ]
        unfixable = ["B"]
        ignore = ["S311"]
